feat: add payment vault smart contract interface
