+++ evmlib/Cargo.toml
@@ -14,7 +14,7 @@ local = []
 external-signer = []
 
 [dependencies]
-alloy = { version = "0.5.3", default-features = false, features = ["std", "reqwest-rustls-tls", "provider-anvil-node", "sol-types", "json", "signers", "contract", "signer-local", "network"] }
+alloy = { version = "0.7.3", default-features = false, features = ["contract", "json-rpc", "network", "node-bindings", "provider-http", "reqwest-rustls-tls", "rpc-client", "rpc-types", "signer-local", "std"] }
 dirs-next = "~2.0.0"
 serde = "=1.0.210"
 serde_with = { version = "3.11.0", features = ["macros"] }
