+++ evmlib/abi/IPaymentVault.json
@@ -1,0 +1,191 @@
+[
+  {
+    "anonymous": false,
+    "inputs": [
+      {
+        "indexed": true,
+        "internalType": "address",
+        "name": "rewardsAddress",
+        "type": "address"
+      },
+      {
+        "indexed": true,
+        "internalType": "uint256",
+        "name": "amount",
+        "type": "uint256"
+      },
+      {
+        "indexed": true,
+        "internalType": "bytes32",
+        "name": "quoteHash",
+        "type": "bytes32"
+      }
+    ],
+    "name": "DataPaymentMade",
+    "type": "event"
+  },
+  {
+    "inputs": [],
+    "name": "AntTokenNull",
+    "type": "error"
+  },
+  {
+    "inputs": [],
+    "name": "BatchLimitExceeded",
+    "type": "error"
+  },
+  {
+    "inputs": [
+      {
+        "components": [
+          {
+            "internalType": "uint256",
+            "name": "closeRecordsStored",
+            "type": "uint256"
+          },
+          {
+            "internalType": "uint256",
+            "name": "maxRecords",
+            "type": "uint256"
+          },
+          {
+            "internalType": "uint256",
+            "name": "receivedPaymentCount",
+            "type": "uint256"
+          },
+          {
+            "internalType": "uint256",
+            "name": "liveTime",
+            "type": "uint256"
+          },
+          {
+            "internalType": "uint256",
+            "name": "networkDensity",
+            "type": "uint256"
+          },
+          {
+            "internalType": "uint256",
+            "name": "networkSize",
+            "type": "uint256"
+          }
+        ],
+        "internalType": "struct IPaymentVault.QuotingMetrics",
+        "name": "_metrics",
+        "type": "tuple"
+      }
+    ],
+    "name": "getQuote",
+    "outputs": [
+      {
+        "internalType": "uint256",
+        "name": "price",
+        "type": "uint256"
+      }
+    ],
+    "stateMutability": "view",
+    "type": "function"
+  },
+  {
+    "inputs": [
+      {
+        "components": [
+          {
+            "internalType": "address",
+            "name": "rewardsAddress",
+            "type": "address"
+          },
+          {
+            "internalType": "uint256",
+            "name": "amount",
+            "type": "uint256"
+          },
+          {
+            "internalType": "bytes32",
+            "name": "quoteHash",
+            "type": "bytes32"
+          }
+        ],
+        "internalType": "struct IPaymentVault.DataPayment[]",
+        "name": "_payments",
+        "type": "tuple[]"
+      }
+    ],
+    "name": "payForQuotes",
+    "outputs": [],
+    "stateMutability": "nonpayable",
+    "type": "function"
+  },
+  {
+    "inputs": [
+      {
+        "components": [
+          {
+            "internalType": "uint256",
+            "name": "closeRecordsStored",
+            "type": "uint256"
+          },
+          {
+            "internalType": "uint256",
+            "name": "maxRecords",
+            "type": "uint256"
+          },
+          {
+            "internalType": "uint256",
+            "name": "receivedPaymentCount",
+            "type": "uint256"
+          },
+          {
+            "internalType": "uint256",
+            "name": "liveTime",
+            "type": "uint256"
+          },
+          {
+            "internalType": "uint256",
+            "name": "networkDensity",
+            "type": "uint256"
+          },
+          {
+            "internalType": "uint256",
+            "name": "networkSize",
+            "type": "uint256"
+          }
+        ],
+        "internalType": "struct IPaymentVault.QuotingMetrics",
+        "name": "_metrics",
+        "type": "tuple"
+      },
+      {
+        "components": [
+          {
+            "internalType": "address",
+            "name": "rewardsAddress",
+            "type": "address"
+          },
+          {
+            "internalType": "uint256",
+            "name": "amount",
+            "type": "uint256"
+          },
+          {
+            "internalType": "bytes32",
+            "name": "quoteHash",
+            "type": "bytes32"
+          }
+        ],
+        "internalType": "struct IPaymentVault.DataPayment",
+        "name": "_payment",
+        "type": "tuple"
+      }
+    ],
+    "name": "verifyPayment",
+    "outputs": [
+      {
+        "internalType": "bool",
+        "name": "isValid",
+        "type": "bool"
+      }
+    ],
+    "stateMutability": "view",
+    "type": "function"
+  }
+]
