--- evmlib/src/contract/data_payments/error.rs
@@ -1,24 +1,0 @@
-// Copyright 2024 MaidSafe.net limited.
-//
-// This SAFE Network Software is licensed to you under The General Public License (GPL), version 3.
-// Unless required by applicable law or agreed to in writing, the SAFE Network Software distributed
-// under the GPL Licence is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-// KIND, either express or implied. Please review the Licences for the specific language governing
-// permissions and limitations relating to use of the SAFE Network Software.
-
-use crate::contract::network_token;
-use alloy::transports::{RpcError, TransportErrorKind};
-
-#[derive(thiserror::Error, Debug)]
-pub enum Error {
-    #[error(transparent)]
-    ContractError(#[from] alloy::contract::Error),
-    #[error(transparent)]
-    RpcError(#[from] RpcError<TransportErrorKind>),
-    #[error(transparent)]
-    NetworkTokenError(#[from] network_token::Error),
-    #[error(transparent)]
-    PendingTransactionError(#[from] alloy::providers::PendingTransactionError),
-    #[error("The transfer limit of 256 has been exceeded")]
-    TransferLimitExceeded,
-}
